requirements.txt
# Core ML
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
accelerate>=0.24.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.1
pydub>=0.25.1
noisereduce>=3.0.0
webrtcvad>=2.0.10

# Transcription
openai-whisper>=20231117

# Post-processing & NLP
language_tool_python>=2.7.1
spacy>=3.7.0

# Evaluation
jiwer>=3.0.3

# Data handling
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0

# Config & logging
pyyaml>=6.0
omegaconf>=2.3.0

# Dev / testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.1.0
